Fix: isOverwrite() not honored while preparing localizations.
When generating the pass package, an error occurred if any localization directory already existed, since the method always tried to create it. This change makes the prepareLocalizations() method behave in the same way as preparePassDirectory(), that is, only try to create the directory if it does not exist, and do nothing if it exists and isOverwrite() is true.